.gemini/antigravity/scratch/pramaan_backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic-settings
google-generativeai
python-multipart
pypdf
jinja2
alembic
fastmcp
duckduckgo-search
streamlit
